+++ src/app/daily/_components/calendar/Calendar.tsx
@@ -1,19 +1,17 @@
 'use client';
 
 import dayjs from 'dayjs';
-import Image from 'next/image';
 import { useRouter } from 'next/navigation';
 import { useState, useMemo } from 'react';
 
-import { WEEK_DAYS } from '@/app/daily/_constants/weekDays';
-import { getCalendarMonthDates, getCalendarWeekDates } from '@/app/daily/_utils/getCalendar';
 import ChevronDown from '@/assets/icons/chevrondown.svg';
 import ChevronUp from '@/assets/icons/chevronup.svg';
 import SolidButton from '@/components/atoms/SolidButton/SolidButton';
-import { DEFAULT_WATERDROP } from '@/constants/image';
-import { formatDateToYYYYMMDD } from '@/utils/formatDay';
+import { WEEK_DAYS } from '@/constants/weekDays';
+import { getCalendarMonthDates, getCalendarWeekDates } from '@/utils/getCalendar';
 
 import * as S from './Calendar.styled';
+import DateCell from './sub/DateCell';
 
 interface CalendarProps {
   streakDateList: string[];
@@ -25,9 +23,7 @@ export default function Calendar({ streakDateList }: CalendarProps) {
   const today = dayjs();
 
   const monthDates = useMemo(() => getCalendarMonthDates(today.year(), today.month() + 1), [today]);
-
   const weekDates = useMemo(() => getCalendarWeekDates(today), [today]);
-
   const dates = isMonthly ? monthDates : weekDates;
 
   return (
@@ -45,28 +41,13 @@ export default function Calendar({ streakDateList }: CalendarProps) {
             {WEEK_DAYS.map((day) => (
               <S.WeekDay key={day}>{day}</S.WeekDay>
             ))}
-            {dates.map((date) => {
-              const streakSet = new Set(streakDateList);
-              const formatted = formatDateToYYYYMMDD(date.toDate());
-              const isAnswered = streakSet.has(formatted);
-              return (
-                <S.DateCell
-                  key={date.toISOString()}
-                  isAnswered={isAnswered}
-                >
-                  {isAnswered ? (
-                    <Image
-                      src={DEFAULT_WATERDROP}
-                      alt='waterdrop'
-                      width={13}
-                      height={16}
-                    />
-                  ) : (
-                    date.date()
-                  )}
-                </S.DateCell>
-              );
-            })}
+            {dates.map((date) => (
+              <DateCell
+                key={date.toISOString()}
+                date={date}
+                streakDateList={streakDateList}
+              />
+            ))}
           </S.Grid>
         </S.CalendarWrapper>
         <SolidButton
